refactor: disallow inline non-test/priv submodules
Disallow inlined submodules those aren't test or private (for sealing
trait implementations) modules in other rust modules. Prefer creating
rust files in a directory instead.
