+++ lyra/src/core/konst/colours.rs
@@ -1,0 +1,4 @@
+pub const EMBED_DEFAULT: u32 = 0x82_6b_d6;
+pub const DOWNVOTE: u32 = 0xdd_2e_44;
+pub const UPVOTE: u32 = 0x58_65_f2;
+pub const POLL_BASE: u32 = 0x00_00_00;
