+++ lyra/src/core/konst/discord.rs
@@ -1,0 +1,3 @@
+pub const BASE_URL: &str = "https://discord.com";
+pub const CDN_URL: &str = "https://cdn.discordapp.com";
+pub const COMMAND_CHOICES_LIMIT: usize = 25;
