+++ lyra/src/core/konst/exit_code.rs
@@ -1,0 +1,18 @@
+/// A harmless notice, confirming something the user might have meant to do
+pub const NOTICE: &str = "❕";
+/// A suspicious notice, implying something the user might not have meant to do
+pub const DUBIOUS: &str = "❔";
+/// A harmless warning
+pub const WARNING: &str = "❗";
+/// Needed information was not found, implying user given an incorrect query
+// pub const NOT_FOUND: &str = "❓";
+/// Invalid command usage, implying unmet conditions
+pub const INVALID: &str = "❌";
+/// User lacked sufficient permissions
+pub const PROHIBITED: &str = "🚫";
+/// Bot lacked sufficient permissions
+pub const FORBIDDEN: &str = "⛔";
+/// Other known errors
+pub const KNOWN_ERROR: &str = "‼️";
+/// Unknown errors
+pub const UNKNOWN_ERROR: &str = "⁉️";
