+++ lyra/src/core/konst/lavaplayer.rs
@@ -1,0 +1,6 @@
+//! Source:
+//! [1](https://github.com/lavalink-devs/lavaplayer/blob/e684e603f0f783d5fcbe1eef9a939b6e9e1cb0e9/main/src/main/java/com/sedmelluq/discord/lavaplayer/track/info/AudioTrackInfoBuilder.java#L14)
+//! [2](https://github.com/lavalink-devs/lavaplayer/blob/e684e603f0f783d5fcbe1eef9a939b6e9e1cb0e9/main/src/main/java/com/sedmelluq/discord/lavaplayer/container/MediaContainerDetection.java#L22)
+
+pub const UNKNOWN_TITLE: &str = "Unknown title";
+pub const UNKNOWN_ARTIST: &str = "Unknown artist";
