+++ lyra/src/core/konst/mod.rs
@@ -1,0 +1,9 @@
+pub mod colours;
+pub mod connection;
+pub mod discord;
+pub mod exit_code;
+pub mod lavaplayer;
+pub mod misc;
+pub mod poll;
+pub mod regex;
+pub mod text;
