+++ lyra/src/core/konst/poll.rs
@@ -1,0 +1,4 @@
+pub const UPVOTE: &str = "🟦";
+pub const DOWNVOTE: &str = "🟥";
+pub const BASE: &str = "⬛";
+pub const RATIO_BAR_SIZE: usize = 16;
