+++ lyra/src/core/statik/mod.rs
@@ -1,0 +1,2 @@
+pub mod application;
+pub mod component;
