+++ lyra/src/error/component/playback/mod.rs
@@ -10,51 +10,6 @@ pub enum HandleVoiceStateUpdateError {
     SetPauseWith(#[from] crate::error::command::require::SetPauseWithError),
 }
 
-pub mod back {
-    use thiserror::Error;
-
-    #[derive(Error, Debug)]
-    #[error(transparent)]
-    pub enum BackError {
-        NotInVoice(#[from] crate::error::NotInVoice),
-        Unsuppressed(#[from] crate::error::command::require::UnsuppressedError),
-        InVoiceWithoutUser(#[from] crate::error::InVoiceWithoutUser),
-        UserOnlyIn(#[from] crate::error::command::check::UserOnlyInError),
-        Respond(#[from] crate::error::core::RespondError),
-        Lavalink(#[from] lavalink_rs::error::LavalinkError),
-    }
-}
-
-pub mod skip {
-    use thiserror::Error;
-
-    #[derive(Error, Debug)]
-    #[error(transparent)]
-    pub enum SkipError {
-        NotPlaying(#[from] crate::error::NotPlaying),
-        NotInVoice(#[from] crate::error::NotInVoice),
-        Unsuppressed(#[from] crate::error::command::require::UnsuppressedError),
-        InVoiceWithoutUser(#[from] crate::error::InVoiceWithoutUser),
-        UsersTrackError(#[from] crate::error::command::check::UsersTrackError),
-        Respond(#[from] crate::error::core::RespondError),
-        Lavalink(#[from] lavalink_rs::error::LavalinkError),
-    }
-}
-
-pub mod play_pause {
-    use thiserror::Error;
-
-    #[derive(Error, Debug)]
-    #[error(transparent)]
-    pub enum Error {
-        NotInVoice(#[from] crate::error::NotInVoice),
-        Unsuppressed(#[from] crate::error::command::require::UnsuppressedError),
-        InVoiceWithoutUser(#[from] crate::error::InVoiceWithoutUser),
-        UserOnlyIn(#[from] crate::error::command::check::UserOnlyInError),
-        NotPlaying(#[from] crate::error::NotPlaying),
-        UsersTrack(#[from] crate::error::command::check::UsersTrackError),
-        Lavalink(#[from] lavalink_rs::error::LavalinkError),
-        Respond(#[from] crate::error::core::RespondError),
-        SetPauseWith(#[from] crate::error::command::require::SetPauseWithError),
-    }
-}
+pub mod back;
+pub mod play_pause;
+pub mod skip;
